Separate testing and performance testing functionality into separate libraries
Only link to testing library for unit and regression tests. Only link to
performance testing library in performance tests.
