+++ libs/pika/algorithms/tests/performance/benchmark_is_heap.cpp
@@ -12,7 +12,6 @@
 #include <pika/modules/timing.hpp>
 #include <pika/parallel/algorithms/generate.hpp>
 #include <pika/parallel/algorithms/is_heap.hpp>
-#include <pika/testing.hpp>
 
 #include <algorithm>
 #include <cstddef>
@@ -212,8 +211,5 @@ int main(int argc, char* argv[])
     init_args.desc_cmdline = desc_commandline;
     init_args.cfg = cfg;
 
-    PIKA_TEST_EQ_MSG(pika::init(pika_main, argc, argv, init_args), 0,
-        "pika main exited with non-zero status");
-
-    return 0;
+    return pika::init(pika_main, argc, argv, init_args);
 }
