+++ libs/pika/algorithms/tests/performance/benchmark_nth_element.cpp
@@ -11,12 +11,11 @@
 #include <pika/assert.hpp>
 #include <pika/init.hpp>
 #include <pika/modules/program_options.hpp>
-#include <pika/testing.hpp>
+#include <pika/parallel/algorithms/nth_element.hpp>
+
+#include <algorithm>
 #include <chrono>
 #include <ciso646>
-
-#include <pika/parallel/algorithms/nth_element.hpp>
-#include <algorithm>
 #include <iostream>
 #include <random>
 #include <stdio.h>
@@ -156,9 +155,5 @@ int main(int argc, char* argv[])
     pika::init_params init_args;
     init_args.cfg = cfg;
 
-    // Initialize and run pika.
-    PIKA_TEST_EQ_MSG(pika::init(pika_main, argc, argv, init_args), 0,
-        "pika main exited with non-zero status");
-
-    return 0;
+    return pika::init(pika_main, argc, argv, init_args);
 }
