+++ libs/pika/async_cuda/tests/performance/cuda_scheduler_throughput.cpp
@@ -26,7 +26,7 @@
 #include <pika/parallel/algorithms/copy.hpp>
 #include <pika/parallel/algorithms/for_each.hpp>
 #include <pika/parallel/algorithms/for_loop.hpp>
-#include <pika/testing.hpp>
+#include <pika/testing/performance.hpp>
 
 #include <algorithm>
 #include <cmath>
@@ -169,10 +169,6 @@ int pika_main(pika::program_options::variables_map& vm)
         matrix_size.uiWC, matrix_size.uiHC);
 
     matrixMultiply<float>(matrix_size, device, iterations);
-
-    // There are no actual tests to run here. We only expect to reach this point
-    // without exceptions having been thrown.
-    PIKA_TEST(true);
 
     return pika::finalize();
 }
