+++ libs/pika/execution/tests/regressions/CMakeLists.txt
@@ -5,13 +5,8 @@
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 set(tests
-    annotated_minmax_2593
-    bulk_then_execute_3182
-    chunk_size_4118
-    future_then_async_executor
-    is_executor_1691
-    parallel_executor_1781
-    split_continuation_clear
+    annotated_minmax_2593 bulk_then_execute_3182 chunk_size_4118
+    future_then_async_executor parallel_executor_1781 split_continuation_clear
 )
 
 if(PIKA_WITH_DATAPAR)
@@ -34,5 +29,17 @@ foreach(test ${tests})
   )
 
   pika_add_regression_test("modules.execution" ${test} ${${test}_PARAMETERS})
-
 endforeach()
+
+if(PIKA_WITH_COMPILE_ONLY_TESTS)
+  set(compile_tests is_executor_1691)
+  foreach(compile_test ${compile_tests})
+    set(sources ${compile_test}.cpp)
+    source_group("Source Files" FILES ${sources})
+    pika_add_regression_compile_test(
+      "modules.execution" ${compile_test}
+      SOURCES ${sources} ${${compile_test}_FLAGS}
+      FOLDER "Tests/Unit/Modules/Execution/CompileOnly"
+    )
+  endforeach()
+endif()
