+++ libs/pika/execution/tests/regressions/parallel_executor_1781.cpp
@@ -4,10 +4,10 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <pika/init.hpp>
-
 #include <pika/algorithm.hpp>
 #include <pika/execution.hpp>
+#include <pika/init.hpp>
+#include <pika/testing.hpp>
 
 #include <vector>
 
@@ -28,5 +28,7 @@ int pika_main()
 
 int main(int argc, char* argv[])
 {
+    // This test should just run without crashing
+    PIKA_TEST(true);
     return pika::init(pika_main, argc, argv);
 }
