+++ testing/performance_testing/include/pika/testing/performance.hpp
@@ -11,6 +11,7 @@
 
 #include <chrono>
 #include <cstddef>
+#include <cstdint>
 #include <map>
 #include <ostream>
 #include <string>
@@ -85,4 +86,6 @@ namespace pika { namespace util {
 
     PIKA_EXPORT void perftests_print_times();
 
+    PIKA_EXPORT void print_cdash_timing(const char* name, double time);
+    PIKA_EXPORT void print_cdash_timing(const char* name, std::uint64_t time);
 }}    // namespace pika::util
