+++ testing/testing/include/pika/testing.hpp
@@ -15,7 +15,6 @@
 #include <pika/preprocessor/expand.hpp>
 #include <pika/preprocessor/nargs.hpp>
 #include <pika/preprocessor/stringize.hpp>
-#include <pika/testing/performance.hpp>
 #include <pika/thread_support/spinlock.hpp>
 #include <pika/util/ios_flags_saver.hpp>
 
@@ -190,9 +189,6 @@ namespace pika { namespace util {
     ////////////////////////////////////////////////////////////////////////////
     PIKA_EXPORT int report_errors();
     PIKA_EXPORT int report_errors(std::ostream& stream);
-    PIKA_EXPORT void print_cdash_timing(const char* name, double time);
-    PIKA_EXPORT void print_cdash_timing(const char* name, std::uint64_t time);
-
 }}    // namespace pika::util
 
 ////////////////////////////////////////////////////////////////////////////////
