+++ tests/performance/local/async_overheads.cpp
@@ -7,7 +7,7 @@
 #include <pika/chrono.hpp>
 #include <pika/future.hpp>
 #include <pika/init.hpp>
-#include <pika/testing.hpp>
+#include <pika/testing/performance.hpp>
 
 #include "worker_timed.hpp"
