+++ tests/performance/local/resume_suspend.cpp
@@ -9,12 +9,11 @@
 // openmp_parallel_region.
 
 #include <pika/chrono.hpp>
-#include <pika/execution_base/this_thread.hpp>
 #include <pika/future.hpp>
 #include <pika/init.hpp>
-#include <pika/testing.hpp>
-
-#include <pika/modules/program_options.hpp>
+#include <pika/program_options.hpp>
+#include <pika/testing/performance.hpp>
+#include <pika/thread.hpp>
 
 #include <cstddef>
 #include <cstdint>
